feat(pdf-editor): resizable-image and scrollable page
